+++ src/engine/task/mod.rs
@@ -5,7 +5,6 @@ pub mod module;
 pub mod module_dag_compiler;
 pub mod module_dag_orchestrator;
 pub mod module_dag_processor;
-pub mod module_node_dag_adapter;
 #[cfg(feature = "store")]
 pub mod repository;
 #[allow(clippy::module_inception)]